+++ process/CHANGELOG.md
@@ -115,3 +115,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Imported process code from `pimalaya-email`.
+
+[1.0.0]: https://crates.io/crates/process-lib/1.0.0
+[0.4.2]: https://crates.io/crates/process-lib/0.4.2
+[0.4.1]: https://crates.io/crates/process-lib/0.4.1
+[0.4.0]: https://crates.io/crates/process-lib/0.4.0
+[0.3.1]: https://crates.io/crates/process-lib/0.3.1
+[0.3.0]: https://crates.io/crates/process-lib/0.3.0
+[0.2.0]: https://crates.io/crates/process-lib/0.2.0
+[0.1.0]: https://crates.io/crates/process-lib/0.1.0
+[0.0.5]: https://crates.io/crates/pimalaya-process/0.0.5
+[0.0.4]: https://crates.io/crates/pimalaya-process/0.0.4
+[0.0.3]: https://crates.io/crates/pimalaya-process/0.0.3
+[0.0.2]: https://crates.io/crates/pimalaya-process/0.0.2
+[0.0.1]: https://crates.io/crates/pimalaya-process/0.0.1